+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Clean doctrings for more consistent style [#215](https://github.com/datagouv/hydra/pull/215)
 - Fix some type hints [#214](https://github.com/datagouv/hydra/pull/214)
 - Add option to force analysis even if resource has not changed[#205](https://github.com/datagouv/hydra/pull/205)
+- Fix get all checks CRUD method [#217](https://github.com/datagouv/hydra/pull/217)
 
 ## 2.0.4 (2024-10-28)